Split funcs into separate submodules
