--- src/operations/mod.rs
@@ -1,2 +1,0 @@
-pub mod shrink_wrap;
-pub mod transform;
